types: Specify non-any type for factory.createFunctionDeclaration
This business of saying `node as ts.FooBar` to follow up on a
`case ts.SyntaxKind.FooBar:` is kind of ugly.  But looking at
the TypeScript implementation itself -- in src/compiler/checker.ts,
for example -- that's exactly what it does all over the place.
So I think there probably isn't a better option.