wip perf-logging in compileHandle
Results:
  2582ms total here in 20 test cases, out of 5000ms total Jest time
         vs. 7500ms total in compileDefinitionString before this change,
         out of 10s total Jest time
  ------
  1975ms in the first `setup`, i.e. in build
     (presumably basically createProgram)
   582ms in `getTypeChecker` -- only half of that the first time,
     then two later ones are >100ms and rest are 0ms
    24ms scattered across `compile`
     1ms residue (from rounding)

Previously it was about 6.0s in createProgram and 1.5s in
getTypeChecker.  So this saved about 4.0s in createProgram and 0.9s in
getTypeChecker.

Possibly a small (~0.3s) but quick win to be had by having the builder
memoize the getTypeChecker result too.  I'd expected TS to handle
that, and it seems like it *mostly* did but not quite.

Then next interesting question: where does the other 2.5s go that's
outside this function?  Prettier?  Flow?  Somewhere else?