wip printers/node [nfc]: Check parent once and for all in getFullyQualifiedName
In each of these functions, if symbol.parent is absent, the control
flow would bypass most of the code to go straight to a particular
`return` statement.  If on the other hand it's present, we'd end up
checking for it repeatedly.

Instead, check for it once up front.  In both cases this significantly
reduces the number of paths one has to consider the control flow
possibly going through when reading this code.