wip printers/node [nfc]: Cut old commented-out code
TODO do this before the symbol.parent early check

This was commented out already when this whole region of code was
first introduced, in 18bab0377 in 2019.  It hasn't been touched
since then.

Most likely it was an experiment that turned out not to be needed,
and the problem it was aimed at was solved some other way.  But even
if not, with no context here about what it was for, whoever tries to
solve the problem in the future will end up having to figure out the
solution themself anyway.  Until then, cut out the commented code,
to make less to read and maintain.