wip printType type references: cut apparently-nop getFQN
TODO pin this down more

As best I can tell, this assignment wasn't doing anything.

First: if it is ever doing anything, it's something we lack a test for.
No tests break when this is removed.  Then, reasoning statically too:

It's assigning to the `escapedText` property on the `typeName` of this
ts.TypeReferenceNode.  The first odd thing about that is that
`typeName` may be a ts.QualifiedName, and in that case it doesn't
ordinarily have an `escapedText` property at all.

But also, there aren't many places we refer to `escapedText` even for
a ts.Identifier.  None of them appear to be downstream of here.

... Hmmmm but that latter reasoning proves too much: the `escapedText`
assignments in smart-identifiers do indeed have an effect, as several
tests prove.