wip [nfc] Always expect type-checker
We always set `checker.current` at the entry points in `compiler.ts`.

Conversely, we already have a number of spots where we unconditionally
assume that `checker.current` is set, e.g. in `src/parse/index.ts` and
`src/printers/declarations.ts`, and we don't seem to be getting
crashes from there -- so there doesn't seem to be some other code path
that can lead to not having it set when we're trying to analyze code.

So, make that assumption everywhere, and simplify the code accordingly.