wip ts: Consistently look at all declarations, not just the first one
If you look at the logic in TS's own src/compiler/checker.ts and
search for references to a "declarations" property, you find that by
far the most common way to examine the declarations of a symbol is to
look through the whole list (modulo short-circuiting, at least) --
whether that's to do something to each one, to see if all of them
satisfy some predicate, or to see if none of them do.

There are a few cases where instead it says something like
`symbol.declarations[0]`, but they're a small minority (perhaps about
10%); and even most of those that do exist are within error-reporting
paths.

So it seems like iterating through the whole list (modulo short
circuits) is the preferred way to do things.  Which makes sense --
that's the sort of thing one has to do to avoid subtle bugs where the
behavior depends in an unintended way on the order in which things
appear in the source code, or the order in which they get processed by
the compiler.

Switch to that.  I suspect that in some or all of these cases, there's
never a real behavior difference -- for example, I'm not sure if it's
possible for a symbol with an EnumMember declaration to have more than
one declaration.  But this style reduces the need to try to think hard
about whether various edge cases like that are possible.

Also add a tiny utility function that helps keep the code simple when
doing this.