wip typeof tests: through a namespace
The current output works, but seems unnecessarily complicated and
unnecessarily divergent from the TS: it says e.g. `typeof n$x.y`.
Why not simply `typeof n.x.y`?  After all, `n` is a value.

More generally: when we need to refer to a value (which I think
corresponds exactly to when we're under a `typeof`, either the
operator or the import modifier), we should be able to refer to
namespaces directly.  TS compiles them to good old JS objects; so the
user's actual runtime code can refer to them via good old
property-access syntax; so in turn our rendering of the TS type
definition must allow that; so we might as well do the same thing
ourselves.