namespaces: Cut bitrotted `flags` check
In current TS, this value of 4098 corresponds to the flag combination
`NodeFlags.Const | NodeFlags.DisallowInContext` -- so, a `const`
variable declaration which was "parsed in a context where
'in-expressions' are not allowed".  That doesn't seem like something
that was really meant here.

The flag values do get reassigned over time -- there's no ABI
stability here.  Before a change in 2019, it would have been
`NodeFlags.Const | NodeFlags.YieldContext`, so "parsed in the 'yield'
context created when parsing a generator".

Until 2016, `NodeFlags.Const | NodeFlags.HasParamDecorators`, or
before that `NodeFlags.Const | NodeFlags.Namespace`, and before that
`NodeFlags.Ambient | NodeFlags.Namespace`.  Before that,
`NodeFlags.Ambient | NodeFlags.DeclarationFile`.  Perhaps that's what
was intended when this line was written?

In any case, this condition hasn't been working for a long time, and
we don't seem to have missed it.  Take it out.