wip namespace nfc stop looking at NodeFlags.Namespace; isIdentifier instead
Having logic that looks at the Namespace flag is confusing because
within TS, it carries no semantic meaning at all: it just says whether
the declaration used the `namespace` keyword or the legacy `module`
keyword in order to say the same thing.  It's consulted only for
printing back the source code (or transformed versions of it), and for
error messages.  So this logic looks like it's a bug.

But within our codebase, we have a transformer early on that sets the
flag for any ModuleDeclaration whose `.name` is an Identifier.  And
conversely, the only ModuleDeclaration nodes where `.name` is not an
Identifier are those from `module 'foo' { … }`, where the keyword must
be `module` and `NodeFlags.Namespace` is never set.  (In the TS
implementation, these come from parseAmbientExternalModuleDeclaration.)
So after that transformer has run, the ModuleDeclaration nodes with
that flag set are exactly those that have an Identifier for their name.

That means that where we check for this flag, we can get exactly the
same result by checking for the name being an Identifier instead --
and that will be closer to a direct expression of what we're really
trying to say there.  So do that instead.