printType [nfc]: Cut use of "first"/"last" marker kinds
The particular kinds of syntax these refer to change over time, pretty
arbitrarily.  They're good for checking whether a given kind is some
sort of literal, or some sort of type node, by checking whether it's
in the range between SyntaxKind.FirstFoo and SyntaxKind.LastFoo.  But
they are not good for this sort of `switch` statement -- they're
basically always going to be a bug in this context.

Right now, the values are:
  FirstTypeNode -> TypePredicate
  LastTypeNode -> ImportType
  FirstLiteralToken -> NumericLiteral

Both TypePredicate and ImportType are covered by cases above these, so
they're in fact no-ops and can just be deleted.  For NumericLiteral,
this case appears to be the one where we actually handle them, so just
change it to the right name.